+++ ranx/__init__.py
@@ -1,6 +1,3 @@
-from .data_structures import Qrels, Run
-from .meta import compare, evaluate, fuse, normalize, optimize_fusion, plot
-
 __all__ = [
     "evaluate",
     "compare",
@@ -11,3 +8,11 @@ __all__ = [
     "Qrels",
     "Run",
 ]
+
+from numba import config
+
+from .data_structures import Qrels, Run
+from .meta import compare, evaluate, fuse, normalize, optimize_fusion, plot
+
+# Set numba threading layer to workqueue
+config.THREADING_LAYER = "workqueue"