+++ ranx/data_structures/run.py
@@ -1,7 +1,7 @@
 import gzip
 import os
 from collections import defaultdict
-from typing import Dict, List
+from typing import Any, Dict, List
 
 import numpy as np
 import pandas as pd
@@ -171,8 +171,24 @@ class Run(object):
             d[q_id] = dict(self[q_id])
         return d
 
+    def to_dataframe(self) -> pd.DataFrame:
+        """Convert Run to Pandas DataFrame with the following columns: `q_id`, `doc_id`, and `score`.
+
+        Returns:
+            pandas.DataFrame: Run as Pandas DataFrame.
+        """
+        data = {"q_id": [], "doc_id": [], "score": []}
+
+        for q_id in self.run:
+            for doc_id in self.run[q_id]:
+                data["q_id"].append(q_id)
+                data["doc_id"].append(doc_id)
+                data["score"].append(self.run[q_id][doc_id])
+
+        return pd.DataFrame.from_dict(data)
+
     def save(self, path: str = "run.json", kind: str = None):
-        """Write `run` to `path` as JSON file, TREC run, or LZ4 file. File type is automatically inferred form the filename extension: ".json" -> "json", ".trec" -> "trec", ".txt" -> "trec", and ".lz4" -> "lz4". Use the "kind" argument to override this behavior.
+        """Write `run` to `path` as JSON file, TREC run, LZ4 file, or Parquet file. File type is automatically inferred form the filename extension: ".json" -> "json", ".trec" -> "trec", ".txt" -> "trec", and ".lz4" -> "lz4", ".parq" -> "parquet", ".parquet" -> "parquet". Use the "kind" argument to override this behavior.
 
         Args:
             path (str, optional): Saving path. Defaults to "run.json".
@@ -189,6 +205,8 @@ class Run(object):
             save_json(self.to_dict(), path)
         elif kind == "lz4":
             save_lz4(self.to_dict(), path)
+        elif kind == "parquet":
+            self.to_dataframe().to_parquet(path, index=False)
         else:
             with open(path, "w") as f:
                 for i, q_id in enumerate(self.run.keys()):
@@ -303,6 +321,35 @@ class Run(object):
         return Run.from_dict(run_py)
 
     @staticmethod
+    def from_parquet(
+        path: str,
+        q_id_col: str = "q_id",
+        doc_id_col: str = "doc_id",
+        score_col: str = "score",
+        pd_kwargs: Dict[str, Any] = None,
+    ):
+        """Convert a Parquet file to ranx.Run.
+
+        Args:
+            path (str): File path.
+            q_id_col (str, optional): Query IDs column. Defaults to "q_id".
+            doc_id_col (str, optional): Document IDs column. Defaults to "doc_id".
+            score_col (str, optional): Relevance scores column. Defaults to "score".
+            pd_kwargs (Dict[str, Any], optional): Additional arguments to pass to `pandas.read_parquet` (see https://pandas.pydata.org/docs/reference/api/pandas.read_parquet.html). Defaults to None.
+
+        Returns:
+            Run: ranx.Run
+        """
+        pd_kwargs = {} if pd_kwargs is None else pd_kwargs
+
+        return Run.from_df(
+            df=pd.read_parquet(path, *pd_kwargs),
+            q_id_col=q_id_col,
+            doc_id_col=doc_id_col,
+            score_col=score_col,
+        )
+
+    @staticmethod
     def from_ranxhub(id: str):
         """Download and load a ranx.Run from ranxhub.
 
@@ -342,6 +389,7 @@ def get_file_kind(path: str = "run.json", kind: str = None) -> str:
     if kind is None:
         kind = os.path.splitext(path)[1][1:]
         kind = "trec" if kind == "txt" else kind
+        kind = "parquet" if kind == "parq" else kind
 
     # Sanity check
     assert kind in {
@@ -349,6 +397,7 @@ def get_file_kind(path: str = "run.json", kind: str = None) -> str:
         "trec",
         "lz4",
         "gz",
-    }, "Error `kind` must be 'json', 'trec', or 'lz4'"
+        "parquet",
+    }, "Error `kind` must be 'json', 'trec', 'lz4', 'gz', or 'parquet'"
 
     return kind