+++ ranx/metrics/bpref.py
@@ -2,12 +2,10 @@ from typing import Union
 
 import numba
 import numpy as np
-from numba import config, njit, prange
+from numba import njit, prange
 
 from .get_hit_lists import _get_hit_list
 from .get_unjudged_lists import _get_unjudged_list
-
-config.THREADING_LAYER = "workqueue"
 
 
 # LOW LEVEL FUNCTIONS ==========================================================