+++ ranx/metrics/rank_biased_precision.py
@@ -2,11 +2,9 @@ from typing import Union
 
 import numba
 import numpy as np
-from numba import config, njit, prange
+from numba import njit, prange
 
 from .common import clean_qrels
-
-config.THREADING_LAYER = "workqueue"
 
 
 # LOW LEVEL FUNCTIONS ==========================================================